code/qarnot/requirements.txt
qarnot
python-dotenv